refactor: update HTTP client, typings, and login method
Refactored the HTTP client, typings, and login query method. The new
method is much simpler than previous one. Additionally, the refactored
HTTP client now includes refresh token behavior under the hood.

Added an invalid credential warning message to the login form.
When a user accesses the page with an expired token, they will be
redirected to the login page and shown a message indicating that the
token has expired and they need to re-login.

Signed-off-by: Muhammad Rizki <kiizuha@gnuweeb.org>
Acked-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20250222225423.1377-6-kiizuha@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>